refactor(stapp2): split enhancements into stapp2_extensions
Move attachment upload, paste-to-upload, sidebar controls, and compact uploader layout into stapp2_extensions.py so stapp2.py stays closer to upstream and PR review is easier.
