+++ frontends/stapp2.py
@@ -1,3 +1,14 @@
+"""
+GenericAgent Streamlit UI (stapp2)
+
+Upstream baseline: lsdefine/GenericAgent frontends/stapp2.py
+Extensions: see stapp2_extensions.py (attachments, sidebar extras, upload UI)
+
+Code review | 审查顺序
+  1. stapp2_extensions.py — attachment & sidebar enhancements
+  2. stapp2.py            — upstream shell + turn expanders (Module F)
+"""
+
 import os, sys
 import html
 if sys.stdout is None: sys.stdout = open(os.devnull, "w")
@@ -18,27 +29,11 @@ import time, json, re, threading, queue
 from datetime import datetime
 from agentmain import GeneraticAgent
 
-# ─── 新增：文件上传 / 粘贴 / 桌面宠物所需额外导入 ─────────────────────────
-import base64, subprocess
-from urllib.request import urlopen
-from urllib.parse import quote
-from io import BytesIO
+from frontends import stapp2_extensions as ext
 
 st.set_page_config(page_title="Cowork", layout="wide")
 
-# ─── 新增：脚本所在目录（用于定位资源文件和桌面宠物脚本） ─────────────────
-script_dir = os.path.dirname(os.path.abspath(__file__))
-
-# ─── 新增：国际化支持（GA_LANG 环境变量，默认中文） ───────────────────────
-LANG = os.environ.get('GA_LANG', 'zh')
-if LANG not in ('zh', 'en'): LANG = 'zh'
-I18N = {
-    'zh': {'force_stop': '强行停止任务', 'reinject_tools': '重新注入System Prompt', 'desktop_pet': '🐱 桌面宠物'},
-    'en': {'force_stop': 'Force Stop',   'reinject_tools': 'Reinject Tools',         'desktop_pet': '🐱 Desktop Pet'},
-}
-def T(key): return I18N.get(LANG, I18N['zh']).get(key, key)
-
-# ─── Anthropic Light Theme CSS（完整版，来自 old 版，保持不变） ──────────
+# ── Upstream: Anthropic light theme CSS | 上游：Anthropic 浅色主题 ─────────
 ANTHROPIC_CSS = """
 <style>
 /* ===== Root variables ===== */
@@ -674,114 +669,7 @@ body:has([data-testid="stSidebar"][aria-expanded="true"]) #sidebar-gear-toggle {
 </style>
 """
 
-# ─── 新增：文件上传区域样式 ──────────────────────────────────────────────
-FILE_UPLOAD_CSS = """
-<style>
-/* 直接隐藏拖放说明区，不动外层 div，避免 overflow 截断按钮 */
-div[data-testid="stFileUploaderDropzoneInstructions"] {
-    display: none !important;
-}
-
-/* section 去掉边框和内边距，仅做弹性容器居中按钮 */
-div[data-testid="stFileUploader"] section {
-    padding: 0 !important;
-    border: none !important;
-    background: transparent !important;
-    display: flex !important;
-    align-items: center !important;
-    justify-content: flex-start !important;
-    min-height: 0 !important;
-}
-
-/* Browse files 按钮直接给定 48×48，不使用绝对定位 */
-div[data-testid="stFileUploader"] section button {
-    width: 48px !important;
-    height: 48px !important;
-    min-width: 48px !important;
-    flex-shrink: 0 !important;
-    padding: 0 !important;
-    margin: 0 !important;
-    border-radius: 12px !important;
-    background: var(--anthropic-bg-secondary) !important;
-    border: 1px solid var(--anthropic-border) !important;
-    color: transparent !important;
-    font-size: 0 !important;
-    position: relative !important;
-    cursor: pointer !important;
-    transition: background 0.2s ease, border-color 0.2s ease !important;
-    box-sizing: border-box !important;
-}
-
-/* + 号绘制在按钮中央 */
-div[data-testid="stFileUploader"] section button::before {
-    content: '+';
-    font-size: 28px;
-    font-weight: 300;
-    line-height: 1;
-    color: var(--anthropic-text-secondary);
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    pointer-events: none;
-}
-
-div[data-testid="stFileUploader"] section button:hover {
-    background: var(--anthropic-primary) !important;
-    border-color: var(--anthropic-primary) !important;
-}
-
-div[data-testid="stFileUploader"] section button:hover::before {
-    color: white;
-}
-
-/* 上传+输入行：禁止换行，上传列固定宽度，输入列填满剩余空间 */
-[data-testid="stHorizontalBlock"]:has([data-testid="stFileUploader"]) {
-    flex-wrap: nowrap !important;
-    align-items: flex-end !important;
-}
-[data-testid="stHorizontalBlock"]:has([data-testid="stFileUploader"]) > *:first-child {
-    flex: 0 0 54px !important;
-    min-width: 54px !important;
-    max-width: 54px !important;
-}
-[data-testid="stHorizontalBlock"]:has([data-testid="stFileUploader"]) > *:last-child {
-    flex: 1 1 0 !important;
-    min-width: 0 !important;
-}
-</style>
-"""
-
-# ─── 新增：粘贴 / 删除信号输入框隐藏（覆盖 ANTHROPIC_CSS 中的定位规则） ──
-# stTextInput 的 header 定位规则来自 ANTHROPIC_CSS，这里专门针对信号输入框覆盖
-PASTE_HIDDEN_INPUT_CSS = """
-<style>
-/* 隐藏粘贴/删除信号输入框及其 Streamlit 包装容器 */
-div[data-testid="stTextInput"]:has(input[placeholder="__paste_image_signal__"]),
-div[data-testid="stTextInput"]:has(input[placeholder="__delete_file_signal__"]) {
-    position: fixed !important;
-    left: -99999px !important;
-    top: -99999px !important;
-    width: 1px !important;
-    height: 1px !important;
-    overflow: hidden !important;
-    opacity: 0 !important;
-    pointer-events: none !important;
-    z-index: -1 !important;
-}
-/* 同时折叠 Streamlit 外层容器，避免占据布局空间 */
-div[data-testid="stElementContainer"]:has(input[placeholder="__paste_image_signal__"]),
-div[data-testid="stElementContainer"]:has(input[placeholder="__delete_file_signal__"]) {
-    height: 0 !important;
-    min-height: 0 !important;
-    overflow: hidden !important;
-    margin: 0 !important;
-    padding: 0 !important;
-}
-</style>
-"""
-
-# ─── Selectbox 宽度修正 JS（来自 old 版，保持不变） ──────────────────────
+# ── Upstream: sidebar selectbox width fix JS | 上游：侧栏下拉框宽度修正 ───
 ANTHROPIC_SELECTBOX_SCRIPT = """
 <div></div>
 <script>
@@ -923,7 +811,7 @@ ANTHROPIC_SELECTBOX_SCRIPT = """
 </script>
 """
 
-# ─── 核心工具函数（来自 old 版，保持不变） ───────────────────────────────
+# ── Upstream: agent init & theme helpers | 上游：Agent 与主题工具 ─────────
 
 @st.cache_resource
 def init():
@@ -1069,262 +957,28 @@ def build_header_agent_badge_script() -> str:
 </script>
 """
 
-# ─── 新增：文件管理工具函数 ──────────────────────────────────────────────
-
-def save_uploaded_file(file_dict: dict) -> str:
-    """保存上传的文件到 temp/uploaded/ 目录，返回绝对路径。"""
-    os.makedirs("temp/uploaded", exist_ok=True)
-    safe_name = re.sub(r"[^A-Za-z0-9._\-一-龥]", "_", file_dict['name'])
-    timestamp = datetime.now().strftime('%Y%m%d_%H%M%S_%f')
-    saved_path = os.path.join("temp", "uploaded", f"{timestamp}_{safe_name}")
-    try:
-        with open(saved_path, "wb") as f:
-            f.write(file_dict['content'])
-        return os.path.abspath(saved_path)
-    except Exception as e:
-        print(f"[ERROR] Failed to save {file_dict['name']}: {e}")
-        return ""
-
-
-def generate_thumbnail(file_dict: dict, size=(80, 80)) -> str:
-    """生成文件缩略图，返回 base64 data URI（图片）或 emoji 图标（其他类型）。"""
-    if file_dict['type'].startswith('image/'):
-        try:
-            from PIL import Image
-            img = Image.open(BytesIO(file_dict['content']))
-            img.thumbnail(size, Image.Resampling.LANCZOS)
-            buf = BytesIO()
-            img.save(buf, format='PNG')
-            b64 = base64.b64encode(buf.getvalue()).decode()
-            return f"data:image/png;base64,{b64}"
-        except Exception as e:
-            print(f"[ERROR] Thumbnail generation failed: {e}")
-            return "📷"
-    ext = os.path.splitext(file_dict['name'])[1].lower()
-    icon_map = {
-        '.pdf': '📄', '.txt': '📝', '.md': '📝', '.doc': '📄', '.docx': '📄',
-        '.xls': '📊', '.xlsx': '📊', '.csv': '📊',
-        '.zip': '📦', '.rar': '📦', '.7z': '📦',
-        '.py': '🐍', '.js': '📜', '.json': '📋', '.xml': '📋',
-        '.mp3': '🎵', '.wav': '🎵', '.mp4': '🎬', '.avi': '🎬',
-    }
-    return icon_map.get(ext, '📎')
-
-
-def build_prompt_with_files(prompt: str, files: list) -> tuple:
-    """构建包含文件附件信息的提示，返回 (agent_prompt, display_prompt)。"""
-    if not files:
-        return prompt, prompt
-
-    attachment_info = ["\n\n[用户上传附件 — 文件已保存到本地磁盘，可用 file_read 工具读取]"]
-    for f in files:
-        saved_path = save_uploaded_file(f)
-        if not saved_path:
-            continue
-        if f['type'].startswith('image/'):
-            b64 = base64.b64encode(f['content']).decode()
-            attachment_info.append(
-                f"\n- [图片附件] {f['name']} ({f['size']} bytes)"
-                f"\n  磁盘路径: {saved_path}"
-                f"\n  data:{f['type']};base64,{b64[:100]}...(truncated)"
-            )
-        elif f['name'].endswith(('.txt', '.md', '.py', '.json', '.log', '.csv', '.yaml', '.yml', '.js', '.ts', '.sql')):
-            try:
-                text = f['content'].decode('utf-8', errors='replace')
-                max_chars = 6000
-                attachment_info.append(
-                    f"\n--- 文本文件: {f['name']} ({f['size']} bytes) ---"
-                    f"\n磁盘路径: {saved_path}"
-                    f"\n{text[:max_chars]}"
-                    + ("\n[内容已截断，请用 file_read 读取完整内容]" if len(text) > max_chars else "")
-                )
-            except Exception:
-                attachment_info.append(f"\n- 文件: {f['name']} (无法解码为文本)\n  磁盘路径: {saved_path}")
-        else:
-            attachment_info.append(f"\n- 文件: {f['name']} ({f['size']} bytes)\n  磁盘路径: {saved_path}")
-
-    agent_prompt = prompt + "\n".join(attachment_info)
-    display_prompt = f"{prompt}\n\n📎 附件: {', '.join(f['name'] for f in files)}"
-    return agent_prompt, display_prompt
-
-
-def render_file_thumbnails():
-    """渲染已上传文件的缩略图，删除按钮通过 JS 写入隐藏信号输入框触发 rerun。"""
-    files = st.session_state.uploaded_files
-    if not files:
-        return
-
-    cards_html = []
-    for idx, f in enumerate(files):
-        thumb = generate_thumbnail(f)
-        safe_name = f['name'].replace('&', '&amp;').replace('"', '&quot;').replace("'", '&#39;').replace('<', '&lt;')
-        name_display = safe_name[:18]
-
-        if thumb.startswith('data:image'):
-            inner = f'<img src="{thumb}" style="width:100%;height:100%;object-fit:cover;display:block;">'
-        else:
-            inner = (
-                f'<div style="width:100%;height:100%;display:flex;'
-                f'align-items:center;justify-content:center;font-size:30px;">'
-                f'{thumb}</div>'
-            )
-
-        card = (
-            f'<div style="position:relative;width:72px;flex-shrink:0;margin-top:6px;">'
-            f'  <div style="width:72px;height:72px;border-radius:8px;overflow:hidden;'
-            f'              border:1px solid #d5cec5;background:#eeece2;">'
-            f'    {inner}'
-            f'  </div>'
-            f'  <div style="font-size:9px;color:#6b7280;text-align:center;margin-top:3px;'
-            f'              white-space:nowrap;overflow:hidden;text-overflow:ellipsis;'
-            f'              max-width:72px;" title="{safe_name}">{name_display}</div>'
-            f'  <div onclick="deleteFile({idx})"'
-            f'       style="position:absolute;top:-6px;right:-6px;width:18px;height:18px;'
-            f'              border-radius:50%;background:#9ca3af;color:white;font-size:13px;'
-            f'              font-weight:bold;display:flex;align-items:center;'
-            f'              justify-content:center;cursor:pointer;line-height:1;'
-            f'              user-select:none;"'
-            f'       onmouseover="this.style.background=\'#ef4444\'"'
-            f'       onmouseout="this.style.background=\'#9ca3af\'">&#215;</div>'
-            f'</div>'
-        )
-        cards_html.append(card)
-
-    n_rows = max(1, (len(files) + 7) // 8)
-    iframe_height = n_rows * 100 + 10
-
-    full_html = (
-        '<!DOCTYPE html><html><head><style>'
-        'html,body{margin:0;padding:0;background:transparent;overflow:hidden;}'
-        '</style></head><body>'
-        '<div style="display:flex;gap:10px;flex-wrap:wrap;'
-        'padding:6px 6px 4px 2px;align-items:flex-start;">'
-        + ''.join(cards_html)
-        + '</div>'
-        '<script>'
-        'function deleteFile(idx){'
-        '  try{'
-        '    var hw=window.parent;'
-        '    var hd=hw.document;'
-        '    var inp=hd.querySelector(\'input[placeholder="__delete_file_signal__"]\');'
-        '    if(!inp)return;'
-        '    var s=Object.getOwnPropertyDescriptor(hw.HTMLInputElement.prototype,"value").set;'
-        '    s.call(inp,String(idx));'
-        '    inp.dispatchEvent(new Event("input",{bubbles:true}));'
-        '    setTimeout(function(){'
-        '      inp.dispatchEvent(new Event("blur",{bubbles:false}));'
-        '      inp.dispatchEvent(new Event("focusout",{bubbles:true}));'
-        '    },100);'
-        '  }catch(e){console.error("deleteFile:",e);}'
-        '}'
-        '</script>'
-        '</body></html>'
-    )
-    _embed_html(full_html, height=iframe_height)
-
-
-@st.dialog("文件预览", width="large")
-def preview_file_dialog():
-    """文件预览模态对话框。"""
-    idx = st.session_state.preview_file_idx
-    if idx is None or idx >= len(st.session_state.uploaded_files):
-        return
-    f = st.session_state.uploaded_files[idx]
-    st.subheader(f"📎 {f['name']}")
-    st.caption(f"类型: {f['type']} | 大小: {f['size']:,} bytes")
-    if f['type'].startswith('image/'):
-        st.image(f['content'], use_container_width=True)
-    elif f['name'].endswith(('.txt', '.md', '.py', '.json', '.log', '.csv', '.yaml', '.yml', '.js', '.ts', '.sql')):
-        try:
-            text = f['content'].decode('utf-8', errors='replace')
-            st.code(text[:5000], language=None)
-            if len(text) > 5000:
-                st.info("内容已截断至前 5000 字符")
-        except Exception as e:
-            st.error(f"无法显示文件内容: {e}")
-    else:
-        st.info("此文件类型不支持预览")
-        st.json({"文件名": f['name'], "类型": f['type'], "大小": f"{f['size']:,} bytes"})
-    if st.button("关闭", use_container_width=True):
-        st.session_state.preview_file_idx = None
-        st.rerun()
-
-
-# ─── 新增：粘贴监听脚本（捕获剪贴板图片，写入隐藏信号输入框触发 rerun） ──
-def build_paste_listener_script() -> str:
-    return """
-<script>
-(function() {
-    var hostWin = window.parent || window;
-    var hostDoc = hostWin.document || document;
-
-    // 幂等：只安装一次粘贴监听
-    if (hostWin.__pasteImageListenerInstalled__) return;
-    hostWin.__pasteImageListenerInstalled__ = true;
-
-    hostDoc.addEventListener('paste', function(e) {
-        var items = e.clipboardData && e.clipboardData.items;
-        if (!items) return;
-        for (var i = 0; i < items.length; i++) {
-            var item = items[i];
-            if (item.type && item.type.startsWith('image/')) {
-                var blob = item.getAsFile();
-                if (!blob) continue;
-                (function(b) {
-                    var reader = new FileReader();
-                    reader.onload = function(ev) {
-                        var b64 = ev.target.result;
-                        var input = hostDoc.querySelector(
-                            'input[placeholder="__paste_image_signal__"]'
-                        );
-                        if (!input) return;
-                        var setter = Object.getOwnPropertyDescriptor(
-                            hostWin.HTMLInputElement.prototype, 'value'
-                        ).set;
-                        setter.call(input, b64);
-                        input.dispatchEvent(new Event('input', { bubbles: true }));
-                        setTimeout(function() {
-                            input.dispatchEvent(new Event('blur', { bubbles: false }));
-                            input.dispatchEvent(new Event('focusout', { bubbles: true }));
-                        }, 100);
-                    };
-                    reader.readAsDataURL(b);
-                })(blob);
-                break;
-            }
-        }
-    });
-})();
-</script>
-"""
-
-# ─── 应用初始化 ──────────────────────────────────────────────────────────
+# ── App bootstrap | 应用启动 ─────────────────────────────────────────────
 
 agent = init()
 
+_UPSTREAM_SESSION_DEFAULTS = {
+    'agent_name': 'GenericAgent', 'streaming': False, 'stopping': False, 'display_queue': None,
+    'partial_response': '', 'reply_ts': '', 'current_prompt': '', 'selected_llm_idx': agent.llm_no,
+    'autonomous_enabled': False, 'messages': [],
+}
+
 def init_session_state():
-    for key, value in {
-        # 原有状态变量（来自 old 版）
-        'agent_name': 'GenericAgent', 'streaming': False, 'stopping': False, 'display_queue': None,
-        'partial_response': '', 'reply_ts': '', 'current_prompt': '', 'selected_llm_idx': agent.llm_no,
-        'autonomous_enabled': False, 'messages': [],
-        # 新增：文件上传状态
-        'uploaded_files': [],       # [{'name', 'type', 'size', 'content'}, ...]
-        'file_uploader_key': 0,     # 用于重置 file_uploader widget
-        'preview_file_idx': None,   # 当前预览的文件索引
-    }.items(): st.session_state.setdefault(key, value)
+    for key, value in _UPSTREAM_SESSION_DEFAULTS.items():
+        st.session_state.setdefault(key, value)
+    ext.register_extension_session_state()
 
 init_session_state()
 
-# 注入主题 CSS 和脚本（来自 old 版，全部恢复）
 st.markdown(ANTHROPIC_CSS, unsafe_allow_html=True)
 st.markdown(build_dynamic_font_css(110.0), unsafe_allow_html=True)
 _embed_html(ANTHROPIC_SELECTBOX_SCRIPT, height=0, width=0)
 _embed_html(build_header_agent_badge_script(), height=0, width=0)
-# 新增：文件上传 & 粘贴支持的 CSS 和脚本
-st.markdown(FILE_UPLOAD_CSS, unsafe_allow_html=True)
-st.markdown(PASTE_HIDDEN_INPUT_CSS, unsafe_allow_html=True)
-_embed_html(build_paste_listener_script(), height=0, width=0)
+ext.inject_extension_assets()
 
 st.session_state.agent_name = 'Generic Agent'
 with st.chat_message("assistant"):
@@ -1332,7 +986,7 @@ with st.chat_message("assistant"):
     st.write("欢迎使用GenericAgent~")
 
 
-# ─── 侧边栏（基于 old 版，新增额外功能按钮） ─────────────────────────────
+# ── Upstream: sidebar (LLM switch) + fork extras | 上游侧栏 + fork 扩展 ──
 @st.fragment
 def render_sidebar():
     llm_options, current_idx = agent.list_llms(), agent.llm_no
@@ -1347,70 +1001,12 @@ def render_sidebar():
         st.toast(f"已切换到备用链路：{llm_labels[selected_idx]}")
         st.rerun()
     st.divider()
-    # 重注入 System Prompt（来自 old 版，新增：额外加载 tool_usable_history.json）
-    if st.button(T('reinject_tools')):
-        agent.llmclient.last_tools = ''
-        try:
-            hist_path = os.path.join(script_dir, '..', 'assets', 'tool_usable_history.json')
-            with open(hist_path, 'r', encoding='utf-8') as f_:
-                tool_hist = json.load(f_)
-            agent.llmclient.backend.history.extend(tool_hist)
-        except Exception:
-            pass
-        st.toast("下次将重新注入System Prompt")
-    # 新增：强行停止任务（不依赖 streaming 状态，随时可用）
-    if st.button(T('force_stop')):
-        agent.abort()
-        st.toast("Stop signal sent")
-        st.rerun()
-    # 新增：桌面宠物启动按钮
-    if st.button(T('desktop_pet')):
-        kwargs = {'creationflags': 0x08} if sys.platform == 'win32' else {}
-        pet_script = os.path.join(script_dir, 'desktop_pet_v2.pyw')
-        if not os.path.exists(pet_script):
-            pet_script = os.path.join(script_dir, 'desktop_pet.pyw')
-        subprocess.Popen([sys.executable, pet_script], **kwargs)
-        # 注入回调：agent 每个 turn 结束后通知桌面宠物显示状态
-        def _pet_req(q):
-            def _do():
-                try: urlopen(f'http://127.0.0.1:41983/?{q}', timeout=2)
-                except Exception: pass
-            threading.Thread(target=_do, daemon=True).start()
-        agent._pet_req = _pet_req
-        if not hasattr(agent, '_turn_end_hooks'):
-            agent._turn_end_hooks = {}
-        def _pet_hook(ctx):
-            parts = [f"Turn {ctx.get('turn', '?')}"]
-            if ctx.get('summary'): parts.append(ctx['summary'])
-            if ctx.get('exit_reason'): parts.append('DONE')
-            _pet_req(f'msg={quote(chr(10).join(parts))}')
-            if ctx.get('exit_reason'): _pet_req('state=idle')
-        agent._turn_end_hooks['pet'] = _pet_hook
-        st.toast("Desktop pet started")
-    # 新增：自主行动控制（仅中文模式显示）
-    if LANG == 'zh':
-        st.divider()
-        if st.button("开始空闲自主行动"):
-            st.session_state.last_reply_time = int(time.time()) - 1800
-            st.toast("已将上次回复时间设为1800秒前")
-            st.rerun()
-        if st.session_state.autonomous_enabled:
-            if st.button("⏸️ 禁止自主行动"):
-                st.session_state.autonomous_enabled = False
-                st.toast("⏸️ 已禁止自主行动")
-                st.rerun()
-            st.caption("🟢 自主行动运行中，会在你离开它30分钟后自动进行")
-        else:
-            if st.button("▶️ 允许自主行动", type="primary"):
-                st.session_state.autonomous_enabled = True
-                st.toast("✅ 已允许自主行动")
-                st.rerun()
-            st.caption("🔴 自主行动已停止")
+    ext.render_sidebar_extras(agent)
 
 with st.sidebar: render_sidebar()
 
 
-# ─── Agent 任务控制（来自 old 版，保持不变） ─────────────────────────────
+# ── Upstream: agent task & chat loop | 上游：任务与对话循环 ───────────────
 
 def start_agent_task(prompt):
     st.session_state.display_queue = agent.put_task(prompt, source="user")
@@ -1443,35 +1039,38 @@ def _get_response_segments(text):
     return [p for p in re.split(r'(?=\*\*LLM Running \(Turn \d+\) \.\.\.\*\*)', text) if p.strip()] or [text]
 
 
-# render_message：新增 intermediate 参数，用于将多轮推理中间段折叠为 expander
-# old 版调用时不传此参数，行为与 old 版完全相同
+_TURN_RUNNING_RE = re.compile(r'^\*\*LLM Running \(Turn \d+\)')
+
+
+def _is_intermediate_turn_segment(segment: str, *, is_last: bool) -> bool:
+    """Fold non-final **LLM Running (Turn N)** blocks into expanders | 将非最终 Turn 块折叠为 expander"""
+    return (not is_last) and bool(_TURN_RUNNING_RE.match(segment.strip()))
+
+
+def _turn_expander_label(content: str) -> str:
+    m = re.match(r'^\*\*LLM Running \(Turn (\d+)\)', content.strip())
+    return f"Turn {m.group(1)} · 推理过程" if m else "推理过程"
+
+
 def render_message(role, content, ts='', unsafe_allow_html=True, intermediate=False):
     with st.chat_message(role):
         if ts: st.markdown(f'<div class="msg-timestamp">{ts}</div>', unsafe_allow_html=True)
         if intermediate:
-            # 中间推理轮次：用 expander 折叠，标签显示 "Turn N · 推理过程"
-            m = re.match(r'^\*\*LLM Running \(Turn (\d+)\)', content.strip())
-            label = f"Turn {m.group(1)} · 推理过程" if m else "推理过程"
-            with st.expander(label, expanded=False):
+            with st.expander(_turn_expander_label(content), expanded=False):
                 st.markdown(content, unsafe_allow_html=unsafe_allow_html)
         else:
             st.markdown(content, unsafe_allow_html=unsafe_allow_html)
 
 
-# finish_streaming_message：新增 intermediate 标记，用于历史消息回放时折叠中间段
 def finish_streaming_message():
     reply_ts = st.session_state.reply_ts
     segments = _get_response_segments(st.session_state.partial_response)
     for i, seg in enumerate(segments):
-        # 非最后一段且格式匹配，才标记为中间推理轮次
-        is_intermediate = (i < len(segments) - 1) and bool(
-            re.match(r'^\*\*LLM Running \(Turn \d+\)', seg.strip())
-        )
         st.session_state.messages.append({
             "role": "assistant",
             "content": seg,
             "time": reply_ts,
-            "intermediate": is_intermediate,
+            "intermediate": _is_intermediate_turn_segment(seg, is_last=(i == len(segments) - 1)),
         })
     st.session_state.last_reply_time = int(time.time())
     st.session_state.partial_response = st.session_state.reply_ts = st.session_state.current_prompt = ''
@@ -1488,103 +1087,32 @@ def render_streaming_area():
         segments = _get_response_segments(st.session_state.partial_response)
         for i, seg in enumerate(segments):
             is_last = (i == len(segments) - 1)
-            is_intermediate = (not is_last) and bool(
-                re.match(r'^\*\*LLM Running \(Turn \d+\)', seg.strip())
-            )
             render_message(
                 "assistant",
                 seg + ("" if not is_last else "▌"),
                 ts=reply_ts,
                 unsafe_allow_html=False,
-                intermediate=is_intermediate,
+                intermediate=_is_intermediate_turn_segment(seg, is_last=is_last),
             )
     if poll_agent_output(): finish_streaming_message()
     else: time.sleep(0.2)
     st.rerun()
 
 
-# ─── 主流程 ──────────────────────────────────────────────────────────────
+# ── Main UI loop | 主界面 ─────────────────────────────────────────────────
 
-# 历史消息渲染（来自 old 版，新增 intermediate 参数传递）
 for msg in st.session_state.messages:
     render_message(msg["role"], msg["content"], ts=msg.get("time", ""),
                    unsafe_allow_html=True, intermediate=msg.get("intermediate", False))
 
-if st.session_state.streaming: render_streaming_area()
+if st.session_state.streaming:
+    render_streaming_area()
 
-# ─── 新增：粘贴图片处理（隐藏信号输入框 + 解码逻辑） ─────────────────────
-_paste_val = st.session_state.get("paste_image_signal", "")
-if _paste_val and _paste_val.startswith("data:image"):
-    try:
-        _header, _b64str = _paste_val.split(",", 1)
-        _mime = _header.split(":")[1].split(";")[0]
-        _ext = _mime.split("/")[1]
-        _content = base64.b64decode(_b64str)
-        if len(_content) <= 10 * 1024 * 1024:
-            _fname = f"pasted_{datetime.now().strftime('%H%M%S%f')[:12]}.{_ext}"
-            st.session_state.uploaded_files.append({
-                'name': _fname, 'type': _mime,
-                'size': len(_content), 'content': _content,
-            })
-    except Exception:
-        pass
-    st.session_state.paste_image_signal = ""
+if ext.process_paste_signal() or ext.process_delete_signal():
     st.rerun()
 
-st.text_input("paste_signal", value="", key="paste_image_signal",
-              label_visibility="collapsed", placeholder="__paste_image_signal__")
+ext.render_signal_inputs()
+prompt = ext.render_attachment_input_row(streaming=st.session_state.streaming)
 
-# ─── 新增：删除文件信号处理 ───────────────────────────────────────────────
-_del_val = st.session_state.get("delete_file_signal", "")
-if _del_val and _del_val.isdigit():
-    idx = int(_del_val)
-    if 0 <= idx < len(st.session_state.uploaded_files):
-        st.session_state.uploaded_files.pop(idx)
-    st.session_state.delete_file_signal = ""
-    st.rerun()
-
-st.text_input("delete_signal", value="", key="delete_file_signal",
-              label_visibility="collapsed", placeholder="__delete_file_signal__")
-
-# ─── 新增：文件缩略图 + 输入区域布局 [文件上传按钮 | 聊天输入框] ──────────
-render_file_thumbnails()
-col_upload, col_input = st.columns([0.08, 0.92])
-
-with col_upload:
-    uploaded_files = st.file_uploader(
-        "上传文件", accept_multiple_files=True, label_visibility="collapsed",
-        key=f"file_uploader_{st.session_state.file_uploader_key}",
-        help="点击上传图片、文档等文件",
-    )
-    if uploaded_files:
-        for uf in uploaded_files:
-            if any(f['name'] == uf.name for f in st.session_state.uploaded_files):
-                continue
-            content = uf.read()
-            if len(content) > 10 * 1024 * 1024:
-                st.warning(f"文件 {uf.name} 超过 10MB，已跳过")
-                continue
-            st.session_state.uploaded_files.append({
-                'name': uf.name,
-                'type': uf.type if uf.type else 'application/octet-stream',
-                'size': len(content), 'content': content,
-            })
-        st.session_state.file_uploader_key += 1
-        st.rerun()
-
-with col_input:
-    prompt = st.chat_input("请输入指令", disabled=st.session_state.streaming)
-
-# 消息发送（来自 old 版，新增文件附件构建和发送后清空逻辑）
 if prompt:
-    files = st.session_state.uploaded_files
-    agent_prompt, display_prompt = build_prompt_with_files(prompt, files)
-    st.session_state.messages.append({
-        "role": "user", "content": display_prompt,
-        "time": datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-    })
-    start_agent_task(agent_prompt)
-    # 新增：发送后清空已上传文件列表
-    st.session_state.uploaded_files.clear()
-    st.session_state.file_uploader_key += 1
-    st.rerun()
+    ext.handle_user_submit(prompt, start_task=start_agent_task)
